% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opmod.R
\name{advancepop}
\alias{advancepop}
\title{advancepop}
\usage{
advancepop(dat, hist, set, hcr, year, verbose = TRUE)
}
\arguments{
\item{dat}{info}

\item{hist}{info}

\item{set}{info}

\item{hcr}{info}

\item{year}{info}

\item{verbose}{info}
}
\description{
advancepop
}
