% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{catch_pope_halfm}
\alias{catch_pope_halfm}
\title{Pope catch equation with half-M convention}
\usage{
catch_pope_halfm(F, M, N)
}
\arguments{
\item{F}{Fishing mortality-at-age for the time step}

\item{M}{Natural mortality-at-age for the time step}

\item{N}{Numbers-at-age at the start of the time step}
}
\value{
Catch-at-age in numbers using half-M timing
}
\description{
Pope catch equation with half-M convention
}
