% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{def.hcr.escape}
\alias{def.hcr.escape}
\title{def.hcr.escape}
\usage{
def.hcr.escape(
  id = "escapement",
  Btarget = "Blim",
  assessmentInterval = 1,
  env = globalenv()
)
}
\description{
def.hcr.escape

Define harvest control rule
}
