% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{def.hcr.pseudo}
\alias{def.hcr.pseudo}
\title{Define harvest control rule with pseudo assessment}
\usage{
def.hcr.pseudo(
  id = "pseudo-msy",
  fractiles = list(catch = 0.5, ffmsy = 0.5, bbmsy = 0.5, bmsy = 0.5, fmsy = 0.5),
  breakpointB = 0,
  clType = "TAC",
  clyears = 1,
  stab = FALSE,
  lower = 0.8,
  upper = 1.2,
  env = globalenv()
)
}
\arguments{
\item{id}{Name/ID of HCR. Default: "pseudo-msy"}

\item{fractiles}{Fractiles. List}

\item{breakpointB}{breakpointb}

\item{clType}{Catch type for uncertainty cap. Default: "TAC".}

\item{clyears}{Number of years for uncertainty cap. Default: 1.}

\item{stab}{Uncertainty cap. Default: FALSE.}

\item{lower}{Upper bound of uncertainty cap. Default: 0.8.}

\item{upper}{Upper bound of uncertainty cap. Default: 1.2.}

\item{env}{Environment. Default: globalenv()}
}
\description{
def.hcr.pseudo
}
