% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{def.hcr.sam}
\alias{def.hcr.sam}
\title{Define harvest control rule for SAM}
\usage{
def.hcr.sam(
  id = "sam",
  nonconvHCR = "conscat",
  silent = TRUE,
  verbose = FALSE,
  env = globalenv()
)
}
\arguments{
\item{id}{Name/ID of HCR. Default: "sam"}

\item{nonconvHCR}{HCR if SAM does not converge. Default: "conscat" (constant catch).}

\item{silent}{silent}

\item{verbose}{verbose}

\item{env}{Environment. Default: globalenv()}
}
\description{
def.hcr.sam
}
