% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{defHCRconscat}
\alias{defHCRconscat}
\title{Define harvest control rule}
\usage{
defHCRconscat(
  id = "conscat",
  constantC = NULL,
  clyears = 1,
  red = NA,
  redyears = 2,
  env = globalenv()
)
}
\description{
Define harvest control rule
}
