% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{defHCRsam}
\alias{defHCRsam}
\title{Define harvest control rule for SAM}
\usage{
defHCRsam(
  id = "sam",
  nonconvHCR = "conscat",
  silent = TRUE,
  verbose = FALSE,
  env = globalenv()
)
}
\arguments{
\item{id}{Name/ID of HCR. Default: "sam"}

\item{nonconvHCR}{HCR if SAM does not converge. Default: "conscat" (constant catch).}

\item{silent}{silent}

\item{verbose}{verbose}

\item{env}{Environment. Default: globalenv()}
}
\description{
Define harvest control rule for SAM
}
