% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{est.metrics}
\alias{est.metrics}
\title{est.metrics}
\usage{
est.metrics(
  mse,
  dat,
  set = NULL,
  mets = "all",
  gen.time = NULL,
  nysim = NULL,
  kobe.probs = c(0.25, 0.5, 0.75)
)
}
\description{
est.metrics
}
