% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{est.ref.levels.stochastic}
\alias{est.ref.levels.stochastic}
\title{est.ref.levels.stochastic}
\usage{
est.ref.levels.stochastic(
  dat,
  set = NULL,
  fmax = 10,
  ncores = parallel::detectCores() - 1,
  ref = c("Fmsy", "Bmsy", "MSY", "B0", "ESBmsy", "SSBmsy", "ESB0"),
  plot = FALSE,
  get.final = FALSE
)
}
\description{
est.ref.levels.stochastic
}
