% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{est.tac}
\alias{est.tac}
\title{est.tac}
\usage{
est.tac(obs., hcr.fun, tacs. = NULL, pars. = NULL)
}
\arguments{
\item{obs.}{observation}

\item{hcr.fun}{HCR function}

\item{tacs.}{tac daa frame}

\item{pars.}{parameters}
}
\description{
est.tac
}
