% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{gen.noise}
\alias{gen.noise}
\title{gen.noise}
\usage{
gen.noise(
  n,
  sd,
  rho = 0,
  bias.cor = 0,
  mv = FALSE,
  dat = NULL,
  by.asmax = FALSE,
  by.length = FALSE,
  hist = NULL
)
}
\arguments{
\item{n}{info}

\item{sd}{info}

\item{rho}{info}

\item{bias.cor}{info}

\item{mv}{info}

\item{dat}{info}

\item{by.asmax}{info}

\item{by.length}{info}

\item{hist}{info}
}
\description{
gen.noise
}
