% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.converged}
\alias{get.converged}
\title{genConvs}
\usage{
get.converged(
  mse,
  convyears = "all",
  convhcrs = "all",
  out = 0,
  verbose = FALSE
)
}
\arguments{
\item{mse}{info}

\item{convyears}{info}

\item{convhcrs}{info}

\item{out}{info}

\item{verbose}{info}
}
\description{
Get converged simulates from a resMSE object
}
