% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.errs}
\alias{get.errs}
\title{get.errs}
\usage{
get.errs(dat, set, x, hist = NULL, rep = FALSE)
}
\arguments{
\item{dat}{info}

\item{set}{info}

\item{x}{info}

\item{hist}{info}

\item{rep}{info}
}
\description{
get.errs
}
