% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.leslie.r}
\alias{get.leslie.r}
\title{get.leslie.r}
\usage{
get.leslie.r(dat)
}
\arguments{
\item{dat}{iamse data set}
}
\description{
Function to estimate intrinsic growth rate r and generation
    time
}
\details{
Function adopted from Henning Winker's spmpriors package
}
