% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.ssb0}
\alias{get.ssb0}
\title{get.ssb0}
\usage{
get.ssb0(
  M,
  mat,
  weight,
  fecun = 1,
  asmax,
  ns,
  spawning,
  R0,
  indage0,
  season,
  FM = NULL
)
}
\arguments{
\item{mat}{maturity ogive}

\item{fecun}{fecundity matrix}

\item{Z}{total mortality}

\item{amax}{number of age classes}
}
\value{
spawning biomass per recruit
}
\description{
Function to calculate SSB (F=0)
}
