% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getMat}
\alias{getMat}
\title{getMat}
\usage{
getMat(Lm50, Lm95, mids, plba)
}
\arguments{
\item{Lm50}{length at 50\% maturity}

\item{Lm95}{length at 95\% maturity}

\item{mids}{midlengths}

\item{plba}{probability of being in mids given age}
}
\description{
Function to estimate maturity at age
}
