% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getMsel}
\alias{getMsel}
\title{getMsel}
\usage{
getMsel(
  Linf,
  K,
  mids,
  plba,
  a = 0.55,
  b = 1.61,
  c = 1.44,
  scale = TRUE,
  lunit = "cm",
  useBelow10 = FALSE
)
}
\arguments{
\item{Linf}{Linf of vBGF}

\item{K}{K of vBGF}

\item{mids}{midlengths}

\item{plba}{probability of being in mids given age}

\item{scale}{logical; scale to have maximum 1 (default: true)}
}
\description{
Function to estimate selectivity of natural mortality
}
