% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getSel}
\alias{getSel}
\title{getSel}
\usage{
getSel(L50, L95, mids, plba)
}
\arguments{
\item{L50}{length at 50\% selectivity}

\item{L95}{length at 95\% selectivity}

\item{mids}{midlengths}

\item{plba}{probability of being in mids given age}
}
\description{
Function to estimate selectivity ogive
}
