% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{gettacs}
\alias{gettacs}
\title{gettacs}
\usage{
gettacs(tacs. = NULL, id. = "", TAC. = NA, obs. = NULL)
}
\arguments{
\item{tacs.}{info}

\item{id.}{info}

\item{TAC.}{info}

\item{obs.}{info}
}
\description{
gettacs
}
