% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamse-package.R
\docType{package}
\name{iamse}
\alias{iamse}
\alias{iamse-package}
\title{iamse: Intra-Annual Management Strategy Evaluation (IAMSE)}
\description{
Provides a framework for intra-annual age-based operating models used in
management strategy evaluation (MSE) of exploited fish stocks.
}
\details{
The package includes tools to:
\itemize{
  \item define and simulate age-based operating models at intra-annual
        time steps,
  \item specify harvest control rules (HCRs) and management procedures,
  \item run management strategy evaluations (MSEs), and
  \item summarise and visualise performance metrics and trade-offs.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tokami.github.io/iamse/}
}

}
\author{
T.K. Mildenberger
}
\keyword{internal}
