% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opmod.R
\name{initpop}
\alias{initpop}
\title{initpop}
\usage{
initpop(dat, set = NULL, out.opt = 1, verbose = TRUE, plot = FALSE, dbg = 0)
}
\arguments{
\item{dat}{info}

\item{set}{info}

\item{out.opt}{info}

\item{verbose}{info}

\item{plot}{info}

\item{dbg}{info}
}
\description{
initpop
}
