% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mclapply.all.os}
\alias{mclapply.all.os}
\title{mclapply.all.os}
\usage{
mclapply.all.os(
  X,
  FUN,
  ...,
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.silent = FALSE,
  mc.cores = getOption("mc.cores", 2L),
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE,
  affinity.list = NULL
)
}
\description{
mclapply comaptible with all OS
}
