% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mclapply.windows}
\alias{mclapply.windows}
\title{mclapply.windows}
\usage{
mclapply.windows(..., mc.cores = parallel::detectCores() - 1)
}
\description{
Alternative parallelisation for windows
}
\details{
Reference: https://www.r-bloggers.com/2014/07/implementing-mclapply-on-windows-a-primer-on-embarrassingly-parallel-computation-on-multicore-systems-with-r/
}
