% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning.R
\name{print.iamse.cond}
\alias{print.iamse.cond}
\title{Print a conditioned operating model}
\usage{
\method{print}{iamse.cond}(x, ...)
}
\arguments{
\item{x}{an object of class `iamse.cond`}

\item{...}{ignored}
}
\value{
Invisibly `x`.
}
\description{
Print a conditioned operating model
}
