% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{repar.sr}
\alias{repar.sr}
\title{repar.sr}
\usage{
repar.sr(dat)
}
\arguments{
\item{dat}{}
}
\value{
spawning biomass per recruit
}
\description{
Reparameterise stock-recruitment relationship
}
\details{
Convert steepness and virbin biomass into other parameterisations.
    So far implemented: h and vb into breakpoint (bp) and R0 for segmented
    regression (hockey-stick)
}
