% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check.dat}
\alias{check.dat}
\title{check.dat}
\usage{
check.dat(dat = NULL, verbose = TRUE)
}
\arguments{
\item{dat}{- data list with all parameters}

\item{verbose}{- print informative messages}
}
\value{
Updated data list
}
\description{
Checks data list and fills missing slots
}
