% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{def.hcr.conscat}
\alias{def.hcr.conscat}
\title{def.hcr.conscat}
\usage{
def.hcr.conscat(
  id = "conscat",
  constantC = NULL,
  clyears = 1,
  red = NA,
  redyears = 2,
  redAlways = FALSE,
  assessmentInterval = 1,
  ffmsySD = 0,
  bbtriggerSD = 0,
  rightRef = 1,
  env = globalenv()
)
}
\description{
def.hcr.conscat

Define harvest control rule
}
