% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{def.hcr.pseudo}
\alias{def.hcr.pseudo}
\title{def.hcr.pseudo}
\usage{
def.hcr.pseudo(
  id = "pseudo-msy",
  fractiles = list(catch = 0.5, ffmsy = 0.5, bbmsy = 0.5, bmsy = 0.5, fmsy = 0.5),
  breakpointB = 0,
  clType = "TAC",
  clyears = 1,
  stab = FALSE,
  lower = 0.8,
  upper = 1.2,
  env = globalenv()
)
}
\arguments{
\item{id}{Name/ID of HCR. Default: "pseudo-msy"}

\item{fractiles}{Fractiles. List}

\item{breakpointB}{breakpointb}

\item{clType}{Catch type for uncertainty cap. Default: "TAC".}

\item{clyears}{Number of years for uncertainty cap. Default: 1.}

\item{stab}{Uncertainty cap. Default: FALSE.}

\item{lower}{Upper bound of uncertainty cap. Default: 0.8.}

\item{upper}{Upper bound of uncertainty cap. Default: 1.2.}

\item{env}{Environment. Default: globalenv()}

\item{safeguard}{safeguard. List}

\item{dteuler}{Time step of Forward Euler scheme.}

\item{reportmode}{Reportmode}

\item{stabilise}{stabilise option of spict. Default: FALSE.}

\item{priorlogn}{prior}

\item{priorlogalpha}{prior}

\item{priorlogbeta}{prior}

\item{priorlogbkfrac}{prior}

\item{fixn}{fixn}

\item{bfac}{bfac}

\item{bref}{bref}

\item{brefType}{brefType}

\item{nyBref}{nyBref}

\item{btar}{btar}

\item{probtar}{probtar}

\item{brule}{Brule}

\item{red}{Reduction}

\item{redyears}{Reduction years}

\item{redAlways}{Reduction always? Default: FALSE.}

\item{rai}{Raising factor. Default: 0.2.}

\item{manstartdY}{Management start year. Default: 0.}

\item{assessmentInterval}{Assessment interval. Default: 1.}

\item{intC}{Interval C. Default: NA.}

\item{nonconvHCR}{HCR if SPiCT does not converge. Default: "conscat" (constant catch).}

\item{bm}{Benchmark. Default: FALSE,  ## e.g. 5 => re-defining Bref at every benchmark}
}
\description{
def.hcr.pseudo

Define harvest control rule with pseudo assessment
}
