% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{def.hcr.ref}
\alias{def.hcr.ref}
\title{def.hcr.ref}
\usage{
def.hcr.ref(consF = 0, fracFmsy = NULL, set. = set, env = globalenv())
}
\arguments{
\item{consF}{either numeric indicating constant F level or "fmsy" for fishing at fmsy}
}
\description{
def.hcr.ref
}
