% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{def.hcr.spict}
\alias{def.hcr.spict}
\title{Define harvest control rule}
\usage{
def.hcr.spict(
  id = "spict-msy",
  fractiles = list(catch = 0.5, ffmsy = 0.5, bbmsy = 0.5, bmsy = 0.5, fmsy = 0.5),
  breakpointB = 0,
  evalBreakpointB = 0,
  safeguard = list(limitB = 0, prob = 0.95),
  dteuler = 1/4,
  reportmode = 1,
  stabilise = 0,
  priorlogn = c(log(2), 2, 1),
  priorlogsdf = c(5, 2, 0),
  priorlogsdc = c(log(0.2), 2, 0),
  priorlogalpha = c(log(1), 2, 1),
  priorlogbeta = c(log(1), 2, 1),
  priorlogbkfrac = c(log(0.5), 2, 0),
  priorlogr = c(log(0.4), 2, 0),
  fixn = NA,
  bfac = NA,
  bref = "current",
  brefType = "target",
  nyBref = 5,
  btar = "bmsy",
  probtar = 0.4,
  brule = 0,
  red = NA,
  redyears = 3,
  redAlways = FALSE,
  rai = 0.2,
  manstartdY = 0,
  assessmentInterval = 1,
  cpvec = FALSE,
  halfAssessInt = FALSE,
  intC = NA,
  nonconvHCR = "conscat",
  clType = "TAC",
  clyears = 1,
  stab = FALSE,
  lower = 0.8,
  upper = 1.2,
  bm = FALSE,
  env = globalenv(),
  scenario = NA
)
}
\arguments{
\item{id}{Name/ID of HCR. Default: "spict-msy"}

\item{fractiles}{Fractiles. List}

\item{breakpointB}{breakpointb}

\item{safeguard}{safeguard. List}

\item{dteuler}{Time step of Forward Euler scheme.}

\item{reportmode}{Reportmode}

\item{stabilise}{stabilise option of spict. Default: FALSE.}

\item{priorlogn}{prior}

\item{priorlogalpha}{prior}

\item{priorlogbeta}{prior}

\item{priorlogbkfrac}{prior}

\item{fixn}{fixn}

\item{bfac}{bfac}

\item{bref}{bref}

\item{brefType}{brefType}

\item{nyBref}{nyBref}

\item{btar}{btar}

\item{probtar}{probtar}

\item{brule}{Brule}

\item{red}{Reduction}

\item{redyears}{Reduction years}

\item{redAlways}{Reduction always? Default: FALSE.}

\item{rai}{Raising factor. Default: 0.2.}

\item{manstartdY}{Management start year. Default: 0.}

\item{assessmentInterval}{Assessment interval. Default: 1.}

\item{intC}{Interval C. Default: NA.}

\item{nonconvHCR}{HCR if SPiCT does not converge. Default: "conscat" (constant catch).}

\item{clType}{Catch type for uncertainty cap. Default: "TAC".}

\item{clyears}{Number of years for uncertainty cap. Default: 1.}

\item{stab}{Uncertainty cap. Default: FALSE.}

\item{lower}{Upper bound of uncertainty cap. Default: 0.8.}

\item{upper}{Upper bound of uncertainty cap. Default: 1.2.}

\item{bm}{Benchmark. Default: FALSE,  ## e.g. 5 => re-defining Bref at every benchmark}

\item{env}{Environment. Default: globalenv()}
}
\description{
def.hcr.spict
}
