% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{def.hcr.ss3}
\alias{def.hcr.ss3}
\title{Define harvest control rule for SS3}
\usage{
def.hcr.ss3(
  id = "ss3",
  exe_dir,
  nonconvHCR = "conscat",
  catch_se = 0.01,
  cpue_se_log = 0.2,
  cpue_month = 6,
  fix_biology = TRUE,
  fix_selectivity = TRUE,
  fix_q = TRUE,
  use_lencomp = TRUE,
  ny_lencomp = 5,
  silent = TRUE,
  debugging_mode = FALSE,
  verbose = TRUE,
  env = globalenv()
)
}
\arguments{
\item{id}{Name/ID of HCR. Default: "ss3"}

\item{nonconvHCR}{HCR if SAM does not converge. Default: "conscat" (constant catch).}

\item{silent}{silent}

\item{verbose}{verbose}

\item{env}{Environment. Default: globalenv()}
}
\description{
def.hcr.ss3
}
