% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrs.R
\name{defHCRconscat}
\alias{defHCRconscat}
\title{Define harvest control rule}
\usage{
defHCRconscat(
  id = "conscat",
  constantC = NULL,
  stab = FALSE,
  lower = 0.8,
  upper = 1.2,
  clyears = 1,
  red = NA,
  redyears = 2,
  env = globalenv()
)
}
\description{
Define harvest control rule
}
