% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{est.cons.mets}
\alias{est.cons.mets}
\title{est.cons.mets}
\usage{
est.cons.mets(mse, dat, mets = "all", sampleSizes = c(0.1, 0.25, 0.5, 0.75, 1))
}
\description{
Estimate consistency of metrics across different sample sizes
}
