% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{est.depletion}
\alias{est.depletion}
\title{est.depletion}
\usage{
est.depletion(
  dat,
  set = NULL,
  fmin = 1e-04,
  fmax = 10,
  nrep = 100,
  verbose = TRUE,
  method = "percentile",
  B0K = FALSE,
  tol = 1e-04,
  do.opt = TRUE
)
}
\description{
est.depletion
}
