% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{est.productivity}
\alias{est.productivity}
\title{est.productivity}
\usage{
est.productivity(
  dat,
  set = NULL,
  ny = 100,
  fmax = 10,
  tsSplit = 8,
  plot = TRUE
)
}
\description{
est.productivity
}
