% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{est.productivity.stochastic}
\alias{est.productivity.stochastic}
\title{est.productivity}
\usage{
est.productivity.stochastic(
  dat,
  set = NULL,
  fmax = 10,
  nf = 1000,
  prob = c(0.1, 0.9),
  ncores = parallel::detectCores() - 1,
  plot = TRUE
)
}
\description{
est.productivity
}
