% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{est.ref.levels}
\alias{est.ref.levels}
\title{est.ref.levels}
\usage{
est.ref.levels(
  dat,
  set = NULL,
  fvec = seq(0, 5, 0.1),
  ncores = parallel::detectCores() - 1,
  ref = c("Fmsy", "Bmsy", "MSY", "ESBmsy", "SSBmsy", "B0"),
  plot = FALSE
)
}
\description{
est.ref.levels
}
