% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{gen.noise}
\alias{gen.noise}
\title{gen.noise}
\usage{
gen.noise(
  n,
  sd,
  rho = 0,
  bias.cor = 0,
  mv = FALSE,
  dat = NULL,
  by.asmax = FALSE,
  by.length = FALSE,
  hist = NULL
)
}
\description{
gen.noise
}
