% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.f}
\alias{get.f}
\title{get.f}
\usage{
get.f(
  target,
  NAA,
  MAA,
  sel,
  weight,
  seasons,
  ns,
  y,
  h,
  asmax,
  mat,
  pzbm,
  spawning,
  R0,
  SR,
  bp,
  recBeta,
  recGamma,
  eR,
  indage0,
  iaFM,
  lastFM = 0.01,
  upper = 100,
  target_cut_off = 0.01,
  type = "catch"
)
}
\description{
get.f
}
\details{
get FM accounting for seasons
}
