% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getM}
\alias{getM}
\title{getM}
\usage{
getM(Linf, K, mids, a = 0.55, b = 1.61, c = 1.44)
}
\arguments{
\item{Linf}{- Linf of vBGF}

\item{K}{- K of vBGF}

\item{mids}{- midlengths}
}
\description{
Function to estimate selectivity of natural mortality
}
