% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getSel}
\alias{getSel}
\title{getSel}
\usage{
getSel(pars, mids, plba, type = "logistic", age = FALSE, ages = NULL)
}
\arguments{
\item{pars}{- parameters, either Ls50 (length at 50\% selectivity) and Ls95
(length at 95\% selectivity) for logistic or LFS, sdl, sdr}

\item{mids}{- midlengths}

\item{plba}{- probability of being in mids given age}

\item{type}{- logistic or dnormal}
}
\description{
Function to estimate selectivity ogive
}
