% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamse-package.R, R/zzz.R
\docType{package}
\name{iamse}
\alias{iamse}
\title{iamse: Intra-annual management strategy evaluation (IAMSE)}
\description{
Contains framework for intra-annual age-based operating model for MSE.

Description of your package
}
\author{
T.K. Mildenberger
}
\keyword{internal}
