% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotiamse.b}
\alias{plotiamse.b}
\title{plotiamse.b}
\usage{
plotiamse.b(
  dat,
  set,
  resMSE,
  yearly = TRUE,
  trendline = TRUE,
  uncert = TRUE,
  med = TRUE,
  hcrs = NA,
  ylim = NULL,
  plot.legend = TRUE,
  col = NA,
  ylab = "Total biomass",
  xlab = "",
  title = "Biomass"
)
}
\description{
plotiamse.b
}
