% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotiamse.cw}
\alias{plotiamse.cw}
\title{plotiamse.cw}
\usage{
plotiamse.cw(
  dat,
  set,
  resMSE,
  yearly = TRUE,
  trendline = TRUE,
  uncert = TRUE,
  med = TRUE,
  hcrs = NA,
  ylim = NULL,
  plot.legend = TRUE,
  col = NA,
  ylab = "Catch",
  xlab = "",
  title = "Catch"
)
}
\description{
plotiamse.cw
}
