% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotiamse.f}
\alias{plotiamse.f}
\title{plotiamse.f}
\usage{
plotiamse.f(
  dat,
  set,
  resMSE,
  yearly = TRUE,
  trendline = TRUE,
  uncert = TRUE,
  med = TRUE,
  hcrs = NA,
  ylim = NULL,
  plot.legend = TRUE,
  col = NA,
  ylab = "Fishing mortality",
  xlab = "",
  title = "Fishing mortality"
)
}
\description{
plotiamse.f
}
