% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotiamse.quant}
\alias{plotiamse.quant}
\title{plotiamse.quant}
\usage{
plotiamse.quant(
  dat,
  set,
  resMSE,
  hcrs = NULL,
  quants = c("bpbmsy.sd", "fmfmsy.sd", "cp.sd"),
  hline = NULL,
  plot.legend = TRUE
)
}
\description{
plotiamse.quant
}
