% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotiamse.tradeoff}
\alias{plotiamse.tradeoff}
\title{plotiamse.tradeoff}
\usage{
plotiamse.tradeoff(
  mets,
  metrics = c("avRelCatch", "PBBlim"),
  hcrs = NA,
  plot.legend = TRUE
)
}
\description{
plotiamse.tradeoff
}
