% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{run.mse}
\alias{run.mse}
\title{run.mse}
\usage{
run.mse(
  dat,
  set,
  fest = NULL,
  ncores = parallel::detectCores() - 1,
  globals = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{data}

\item{set}{settings}

\item{fest}{F}

\item{ncores}{for parallel}

\item{globals}{list with extra objects to be passed on to cluseters (only if
ncores > 1)}

\item{verbose}{logical; print info}
}
\value{
a list with MSE results
}
\description{
Run MSE
}
