% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{stocklist}
\alias{stocklist}
\title{Fisheries data included in Polacheck et al. (1993).}
\format{Data are lists containing data}
\usage{
data(stocklist)
}
\description{
Fisheries data included in Polacheck et al. (1993).
}
\details{
Fisheries data for south Atlantic albacore, northern Namibian hake, and New Zealand rock lobster.
}
\keyword{datasets}
